+++ Samples/Space/Scripts/ReturnToMenuButton.cs
@@ -1,0 +1,11 @@
+﻿using Godot;
+
+namespace YarnSpinnerGodot.Samples;
+
+public partial class ReturnToMenuButton : Button
+{
+    public override void _Ready()
+    {
+        Pressed += SampleEntryPoint.Return;
+    }
+}